refactor(sandbox): move spawn helpers (classifyEntityId, currentEnemyId, spawnWithCount) to sandbox_spawn.js and import into sandbox_panel.js
